moved more of the app to server components
