+++ components/ui/forms/departure-form.tsx
@@ -1,6 +1,5 @@
 'use client'
 import { useAddRentalStart } from "@/mutations/mutations";
-import MainContainer from "../containers/main-container";
 import Inputs from "./inputs";
 import { ItemTypes } from "@/types/types";
 import { Suspense } from "react";
@@ -8,7 +7,6 @@ import InputDepartureFormFallback from "../fallbacks/input-departure-form-fallba
 
 export default function DepartureForm({ itemTypesPromise }: { itemTypesPromise: Promise<ItemTypes[]> }) {
     const { mutate, isPending, isError } = useAddRentalStart();
-
     return (
         <>
             <h1 className="text-2xl">Show a guests departure </h1>
